Use integer for price column

+++ tables.sql
@@ -2,7 +2,7 @@ create table properties (
     id bigint not null primary key,
     location geography(Point) not null,
     address text,
-    price smallint not null,
+    price integer not null,
     bedrooms smallint,
     bathrooms smallint
 );
